pyproject.toml
[project]
name = "user-svc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "pytest>=8.3.3",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.36",
    "typer>=0.12.5",
    "uvicorn[standard]>=0.31.0",
]
